fix simple usage of all of

+++ test/Acceptance/ExpectedClasses/Schema/Test5.php
@@ -12,7 +12,7 @@ readonly class Test5 implements JsonSerializable
     public function __construct(
         public bool $ok,
         public Test1 $test,
-        public int $money,
+        public float $money,
         public DateTimeInterface $date,
         public string $fullName,
         public ?string $address,
